+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Remove the deprecated `hide_user_not_found` configuration option, use `expose_security_errors` instead
  * Remove `LazyFirewallContext::__invoke()`
+ * Make `ExpressionCacheWarmer` class `final`
 
 7.4
 ---